some api cleanup
